Add conf level saving, Extend dashboardUtil
-> Add configuration saving level
-> Remove group generic functions to conf level (DashboardConfigService)
-> Work on dashboard util (fix computeAdvancement and add new generic methods for util)
-> Work on  Campaign_evolution indicator

+++ source/src/main/java/org/cerberus/crud/dao/impl/DashboardEntryDAO.java
@@ -58,7 +58,7 @@ public class DashboardEntryDAO implements IDashboardEntryDAO {
         try {
             Connection connection = databaseSpring.connect();
             PreparedStatement preStat = connection.prepareStatement(query.toString());
-            preStat.setInt(1, dashboardgroup.getId());
+            preStat.setLong(1, dashboardgroup.getId());
             ResultSet rs = preStat.executeQuery();
             while (rs.next()) {
                 response.add(this.loadFromResultSet(rs));
